requirements.txt
# TensorFlow 2.15.0 (compatible with Python 3.11 and Object Detection API)
tensorflow==2.15.0

# Object Detection API Dependencies
protobuf>=3.20.0
pillow
lxml
matplotlib
Cython
contextlib2
tf-slim
pycocotools
opencv-python

# Data Processing
pandas
numpy

# Visualization
tensorboard
